Creepy Zombie Hand icon
I don't know if you overhead my conversation with one of our classmates
on Monday, but he suggested a zombie icon for our map, so I made one.
I'm thinking there definitely might be some good zombie references for
the zombie enthusiasts. There's still one more icon to be made that's
unaccounted for, it's the forgotten landmarks icon. I'll get on that
soon. In the meantime, here's the icon. Feedback welcome.